requirements.txt
google-generativeai>=0.7.0
python-dotenv>=1.0.1
rapidfuzz>=3.9.6
pandas>=2.2.2
numpy>=2.0.0
scikit-learn>=1.5.0
requests>=2.31.0

# Ship-fee feature deps
fastapi>=0.111.0
uvicorn>=0.30.0
redis>=5.0.0

